@@ -10,8 +10,8 @@
     "orbit-db": "^0.19.9",
     "prop-types": "^15.6.1",
     "ramda": "^0.25.0",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
+    "react": "^16.7.0-alpha.0",
+    "react-dom": "^16.7.0-alpha.0",
     "react-router-dom": "^4.2.2",
     "react-scripts": "2.0.0-next.a671462c",
     "rxjs": "^5.5.6",
